Refactored handlers and play. Some things are messy because of imports from drummachine.
