+++ public/js/drummachine.js
@@ -1,39 +1,33 @@
 const beatMod = require('./beat')
 const synthMod = require('./synthesis')
-const drawMod = require('./draw.js')
+const drawMod = require('./draw')
 const kitMod = require('./kit')
 const impulseMod = require('./impulse')
+const handlersMod = require('./handlers')
+const playMod = require('./play')
 
 // Events
 // init() once the page has finished loading.
 //window.onload = init;
 
-var timerWorker = null; // Worker thread to send us scheduling messages.
-var context; exports.context = context;
-var convolver;
-var compressor;
-var masterGainNode;
-var effectLevelNode;
-var filterNode;
 
-// Each effect impulse response has a specific overall desired dry and wet volume.
-// For example in the telephone filter, it's necessary to make the dry volume 0 to correctly hear the effect.
-var effectDryMix = 1.0;
-var effectWetMix = 1.0;
+var context; exports.context = context;
+var convolver; exports.convolver = convolver;
+var compressor; exports.compressor = compressor;
+var masterGainNode; exports.masterGainNode = masterGainNode;
+var effectLevelNode; exports.effectLevelNode = effectLevelNode;
+var filterNode; exports.filterNode = filterNode;
+
 
 var timeoutId;
 
-var startTime;
 
 
-var kMaxSwing = .08;
 
 
 
 kickPitch = snarePitch = hihatPitch = tom1Pitch = tom2Pitch = tom3Pitch = 0;
 
-var mouseCapture = null;
-var mouseCaptureOffset = 0;
 
 var noteTime = 0.0;
 
@@ -185,14 +179,20 @@ exports.initDrums = function(cmInstance) {
     convolver = context.createConvolver();
     convolver.connect(effectLevelNode);
 
+    exports.filterNode = filterNode;
+    exports.finalMixNode = finalMixNode;
+    exports.compressor = compressor;
+    exports.masterGainNode = masterGainNode;
+    exports.effectLevelNode = effectLevelNode;
+    exports.convolver = convolver;
 
     var elKitCombo = document.getElementById('kitcombo');
-    elKitCombo.addEventListener("mousedown", handleKitComboMouseDown, true);
+    elKitCombo.addEventListener("mousedown", handlersMod.handleKitComboMouseDown, true);
 
     var elEffectCombo = document.getElementById('effectcombo');
-    elEffectCombo.addEventListener("mousedown", handleEffectComboMouseDown, true);
+    elEffectCombo.addEventListener("mousedown", handlersMod.handleEffectComboMouseDown, true);
 
-    document.body.addEventListener("mousedown", handleBodyMouseDown, true);
+    document.body.addEventListener("mousedown", handlersMod.handleBodyMouseDown, true);
 
     initControls();
     updateControls();
@@ -203,11 +203,13 @@ exports.initDrums = function(cmInstance) {
     // Obtain a blob URL reference to our worker 'file'.
     var timerWorkerBlobURL = window.URL.createObjectURL(timerWorkerBlob);
 
-    timerWorker = new Worker(timerWorkerBlobURL);
-    timerWorker.onmessage = function(e) {
-      schedule();
+    tmp = new Worker(timerWorkerBlobURL);
+    tmp.onmessage = function(e) {
+      playMod.schedule();
     };
-    timerWorker.postMessage('init'); // Start the worker.
+    tmp.postMessage('init'); // Start the worker.
+
+    handlersMod.setTimerWorker(tmp)
 
 }
 
@@ -218,42 +220,42 @@ function initControls() {
     makeEffectList();
 
     // sliders
-    document.getElementById('effect_thumb').addEventListener('mousedown', handleSliderMouseDown, true);
-    document.getElementById('tom1_thumb').addEventListener('mousedown', handleSliderMouseDown, true);
-    document.getElementById('tom2_thumb').addEventListener('mousedown', handleSliderMouseDown, true);
-    document.getElementById('tom3_thumb').addEventListener('mousedown', handleSliderMouseDown, true);
-    document.getElementById('hihat_thumb').addEventListener('mousedown', handleSliderMouseDown, true);
-    document.getElementById('snare_thumb').addEventListener('mousedown', handleSliderMouseDown, true);
-    document.getElementById('kick_thumb').addEventListener('mousedown', handleSliderMouseDown, true);
-    document.getElementById('swing_thumb').addEventListener('mousedown', handleSliderMouseDown, true);
+    document.getElementById('effect_thumb').addEventListener('mousedown', handlersMod.handleSliderMouseDown, true);
+    document.getElementById('tom1_thumb').addEventListener('mousedown', handlersMod.handleSliderMouseDown, true);
+    document.getElementById('tom2_thumb').addEventListener('mousedown', handlersMod.handleSliderMouseDown, true);
+    document.getElementById('tom3_thumb').addEventListener('mousedown', handlersMod.handleSliderMouseDown, true);
+    document.getElementById('hihat_thumb').addEventListener('mousedown', handlersMod.handleSliderMouseDown, true);
+    document.getElementById('snare_thumb').addEventListener('mousedown', handlersMod.handleSliderMouseDown, true);
+    document.getElementById('kick_thumb').addEventListener('mousedown', handlersMod.handleSliderMouseDown, true);
+    document.getElementById('swing_thumb').addEventListener('mousedown', handlersMod.handleSliderMouseDown, true);
 
-    document.getElementById('effect_thumb').addEventListener('dblclick', handleSliderDoubleClick, true);
-    document.getElementById('tom1_thumb').addEventListener('dblclick', handleSliderDoubleClick, true);
-    document.getElementById('tom2_thumb').addEventListener('dblclick', handleSliderDoubleClick, true);
-    document.getElementById('tom3_thumb').addEventListener('dblclick', handleSliderDoubleClick, true);
-    document.getElementById('hihat_thumb').addEventListener('dblclick', handleSliderDoubleClick, true);
-    document.getElementById('snare_thumb').addEventListener('dblclick', handleSliderDoubleClick, true);
-    document.getElementById('kick_thumb').addEventListener('dblclick', handleSliderDoubleClick, true);
-    document.getElementById('swing_thumb').addEventListener('dblclick', handleSliderDoubleClick, true);
+    document.getElementById('effect_thumb').addEventListener('dblclick', handlersMod.handleSliderDoubleClick, true);
+    document.getElementById('tom1_thumb').addEventListener('dblclick', handlersMod.handleSliderDoubleClick, true);
+    document.getElementById('tom2_thumb').addEventListener('dblclick', handlersMod.handleSliderDoubleClick, true);
+    document.getElementById('tom3_thumb').addEventListener('dblclick', handlersMod.handleSliderDoubleClick, true);
+    document.getElementById('hihat_thumb').addEventListener('dblclick', handlersMod.handleSliderDoubleClick, true);
+    document.getElementById('snare_thumb').addEventListener('dblclick', handlersMod.handleSliderDoubleClick, true);
+    document.getElementById('kick_thumb').addEventListener('dblclick', handlersMod.handleSliderDoubleClick, true);
+    document.getElementById('swing_thumb').addEventListener('dblclick', handlersMod.handleSliderDoubleClick, true);
 
     // tool buttons
-    document.getElementById('play').addEventListener('mousedown', handlePlay, true);
-    document.getElementById('stop').addEventListener('mousedown', handleStop, true);
-    document.getElementById('save').addEventListener('mousedown', handleSave, true);
-    document.getElementById('save_ok').addEventListener('mousedown', handleSaveOk, true);
-    document.getElementById('load').addEventListener('mousedown', handleLoad, true);
-    document.getElementById('load_ok').addEventListener('mousedown', handleLoadOk, true);
-    document.getElementById('load_cancel').addEventListener('mousedown', handleLoadCancel, true);
-    document.getElementById('reset').addEventListener('mousedown', handleReset, true);
-    document.getElementById('demo1').addEventListener('mousedown', handleDemoMouseDown, true);
-    document.getElementById('demo2').addEventListener('mousedown', handleDemoMouseDown, true);
-    document.getElementById('demo3').addEventListener('mousedown', handleDemoMouseDown, true);
-    document.getElementById('demo4').addEventListener('mousedown', handleDemoMouseDown, true);
-    document.getElementById('demo5').addEventListener('mousedown', handleDemoMouseDown, true);
+    document.getElementById('play').addEventListener('mousedown', handlersMod.handlePlay, true);
+    document.getElementById('stop').addEventListener('mousedown', handlersMod.handleStop, true);
+    document.getElementById('save').addEventListener('mousedown', handlersMod.handleSave, true);
+    document.getElementById('save_ok').addEventListener('mousedown', handlersMod.handleSaveOk, true);
+    document.getElementById('load').addEventListener('mousedown', handlersMod.handleLoad, true);
+    document.getElementById('load_ok').addEventListener('mousedown', handlersMod.handleLoadOk, true);
+    document.getElementById('load_cancel').addEventListener('mousedown', handlersMod.handleLoadCancel, true);
+    document.getElementById('reset').addEventListener('mousedown', handlersMod.handleReset, true);
+    document.getElementById('demo1').addEventListener('mousedown', handlersMod.handleDemoMouseDown, true);
+    document.getElementById('demo2').addEventListener('mousedown', handlersMod.handleDemoMouseDown, true);
+    document.getElementById('demo3').addEventListener('mousedown', handlersMod.handleDemoMouseDown, true);
+    document.getElementById('demo4').addEventListener('mousedown', handlersMod.handleDemoMouseDown, true);
+    document.getElementById('demo5').addEventListener('mousedown', handlersMod.handleDemoMouseDown, true);
 
     var elBody = document.getElementById('body');
-    elBody.addEventListener('mousemove', handleMouseMove, true);
-    elBody.addEventListener('mouseup', handleMouseUp, true);
+    elBody.addEventListener('mousemove', handlersMod.handleMouseMove, true);
+    elBody.addEventListener('mouseup', handlersMod.handleMouseUp, true);
 
     document.getElementById('tempoinc').addEventListener('mousedown', beatMod.tempoIncrease, true);
     document.getElementById('tempodec').addEventListener('mousedown', beatMod.tempoDecrease, true);
@@ -265,7 +267,7 @@ function initButtons() {
     for (i = 0; i < beatMod.loopLength; ++i) {
         for (j = 0; j < kitMod.kNumInstruments; j++) {
                 elButton = document.getElementById(kitMod.instruments[j] + '_' + i);
-                elButton.addEventListener("mousedown", handleButtonMouseDown, true);
+                elButton.addEventListener("mousedown", handlersMod.handleButtonMouseDown, true);
         }
     }
 }
@@ -277,13 +279,13 @@ function makeEffectList() {
 
     var elItem = document.createElement('li');
     elItem.innerHTML = 'None';
-    elItem.addEventListener("mousedown", handleEffectMouseDown, true);
+    elItem.addEventListener("mousedown", handlersMod.handleEffectMouseDown, true);
 
     for (var i = 0; i < numEffects; i++) {
         var elItem = document.createElement('li');
         elItem.innerHTML = impulseMod.impulseResponseInfoList[i].name;
         elList.appendChild(elItem);
-        elItem.addEventListener("mousedown", handleEffectMouseDown, true);
+        elItem.addEventListener("mousedown", handlersMod.handleEffectMouseDown, true);
     }
 }
 
@@ -295,396 +297,15 @@ function makeKitList() {
         var elItem = document.createElement('li');
         elItem.innerHTML = kitMod.kitNamePretty[i];
         elList.appendChild(elItem);
-        elItem.addEventListener("mousedown", handleKitMouseDown, true);
-    }
-}
-
-
-function advanceNote() {
-
-    synthMod.updatePatternFromCode();
-
-    // Advance time by a 16th note...
-    var secondsPerBeat = 60.0 / beatMod.theBeat.tempo;
-
-    beatMod.setRhythmIndex(beatMod.rhythmIndex + 1);
-    if (beatMod.rhythmIndex == beatMod.loopLength) {
-        beatMod.setRhythmIndex(0);
-    }
-
-        // apply swing
-    if (beatMod.rhythmIndex % 2) {
-        noteTime += (0.25 + kMaxSwing * beatMod.theBeat.swingFactor) * secondsPerBeat;
-    } else {
-        noteTime += (0.25 - kMaxSwing * beatMod.theBeat.swingFactor) * secondsPerBeat;
-    }
-}
-
-function playNote(buffer, pan, x, y, z, sendGain, mainGain, playbackRate, noteTime) {
-    // Create the note
-    var voice = context.createBufferSource();
-    voice.buffer = buffer;
-    voice.playbackRate.value = playbackRate;
-
-    // Optionally, connect to a panner
-    var finalNode;
-    if (pan) {
-        var panner = context.createPanner();
-        panner.panningModel = "HRTF";
-        panner.setPosition(x, y, z);
-        voice.connect(panner);
-        finalNode = panner;
-    } else {
-        finalNode = voice;
-    }
-
-    // Connect to dry mix
-    var dryGainNode = context.createGain();
-    dryGainNode.gain.value = mainGain * effectDryMix;
-    finalNode.connect(dryGainNode);
-    dryGainNode.connect(masterGainNode);
-
-    // Connect to wet mix
-    var wetGainNode = context.createGain();
-    wetGainNode.gain.value = sendGain;
-    finalNode.connect(wetGainNode);
-    wetGainNode.connect(convolver);
-
-    voice.start(noteTime);
-}
-
-function schedule() {
-    var currentTime = context.currentTime;
-
-    // The sequence starts at startTime, so normalize currentTime so that it's 0 at the start of the sequence.
-    currentTime -= startTime;
-
-    while (noteTime < currentTime + 0.120) {
-        // Convert noteTime to context time.
-        var contextPlayTime = noteTime + startTime;
-
-        // Kick
-        if (beatMod.theBeat.rhythm1[beatMod.rhythmIndex] && instrumentActive[0]) {
-            playNote(kitMod.currentKit.kickBuffer, false, 0,0,-2, 0.5, kitMod.volumes[beatMod.theBeat.rhythm1[beatMod.rhythmIndex]] * 1.0, kickPitch, contextPlayTime);
-        }
-
-        // Snare
-        if (beatMod.theBeat.rhythm2[beatMod.rhythmIndex] && instrumentActive[1]) {
-            playNote(kitMod.currentKit.snareBuffer, false, 0,0,-2, 1, kitMod.volumes[beatMod.theBeat.rhythm2[beatMod.rhythmIndex]] * 0.6, snarePitch, contextPlayTime);
-        }
-
-        // Hihat
-        if (beatMod.theBeat.rhythm3[beatMod.rhythmIndex] && instrumentActive[2]) {
-            // Pan the hihat according to sequence position.
-            playNote(kitMod.currentKit.hihatBuffer, true, 0.5*beatMod.rhythmIndex - 4, 0, -1.0, 1, kitMod.volumes[beatMod.theBeat.rhythm3[beatMod.rhythmIndex]] * 0.7, hihatPitch, contextPlayTime);
-        }
-
-        // Toms
-        if (beatMod.theBeat.rhythm4[beatMod.rhythmIndex] && instrumentActive[3]) {
-            playNote(kitMod.currentKit.tom1, false, 0,0,-2, 1, kitMod.volumes[beatMod.theBeat.rhythm4[beatMod.rhythmIndex]] * 0.6, tom1Pitch, contextPlayTime);
-        }
-
-        if (beatMod.theBeat.rhythm5[beatMod.rhythmIndex] && instrumentActive[4]) {
-            playNote(kitMod.currentKit.tom2, false, 0,0,-2, 1, kitMod.volumes[beatMod.theBeat.rhythm5[beatMod.rhythmIndex]] * 0.6, tom2Pitch, contextPlayTime);
-        }
-
-        if (beatMod.theBeat.rhythm6[beatMod.rhythmIndex] && instrumentActive[5]) {
-            playNote(kitMod.currentKit.tom3, false, 0,0,-2, 1, kitMod.volumes[beatMod.theBeat.rhythm6[beatMod.rhythmIndex]] * 0.6, tom3Pitch, contextPlayTime);
-        }
-
-
-        // Attempt to synchronize drawing time with sound
-        if (noteTime != drawMod.lastDrawTime) {
-            drawMod.setLastDrawTime(noteTime);
-            drawMod.drawPlayhead((beatMod.rhythmIndex + 15) % 16);
-        }
-
-        advanceNote();
-    }
-}
-
-function playDrum(noteNumber, velocity) {
-    switch (noteNumber) {
-        case 0x24:
-            playNote(kitMod.currentKit.kickBuffer,  false, 0,0,-2,  0.5, (velocity / 127), kickPitch,  0);
-            break;
-        case 0x26:
-            playNote(kitMod.currentKit.snareBuffer, false, 0,0,-2,  1,   (velocity / 127), snarePitch, 0);
-            break;
-        case 0x28:
-            playNote(kitMod.currentKit.hihatBuffer, true,  0,0,-1.0,1,   (velocity / 127), hihatPitch, 0);
-            break;
-        case 0x2d:
-            playNote(kitMod.currentKit.tom1,        false, 0,0,-2,  1,   (velocity / 127), tom1Pitch,  0);
-            break;
-        case 0x2f:
-            playNote(kitMod.currentKit.tom2,        false, 0,0,-2,  1,   (velocity / 127), tom2Pitch,  0);
-            break;
-        case 0x32:
-            playNote(kitMod.currentKit.tom3,        false, 0,0,-2,  1,   (velocity / 127), tom3Pitch,  0);
-            break;
-        default:
-            console.log("note:0x" + noteNumber.toString(16) );
+        elItem.addEventListener("mousedown", handlersMod.handleKitMouseDown, true);
     }
 }
 
 
 
 
-function handleSliderMouseDown(event) {
-    mouseCapture = event.target.id;
 
-    // calculate offset of mousedown on slider
-    var el = event.target;
-    if (mouseCapture == 'swing_thumb') {
-        var thumbX = 0;
-        do {
-            thumbX += el.offsetLeft;
-        } while (el = el.offsetParent);
 
-        mouseCaptureOffset = event.pageX - thumbX;
-    } else {
-        var thumbY = 0;
-        do {
-            thumbY += el.offsetTop;
-        } while (el = el.offsetParent);
-
-        mouseCaptureOffset = event.pageY - thumbY;
-    }
-}
-
-function handleSliderDoubleClick(event) {
-    var id = event.target.id;
-    if (id != 'swing_thumb' && id != 'effect_thumb') {
-        mouseCapture = null;
-        sliderSetValue(event.target.id, 0.5);
-        updateControls();
-    }
-}
-
-function handleMouseMove(event) {
-    if (!mouseCapture) return;
-
-    var elThumb = document.getElementById(mouseCapture);
-    var elTrack = elThumb.parentNode;
-
-    if (mouseCapture != 'swing_thumb') {
-        var thumbH = elThumb.clientHeight;
-        var trackH = elTrack.clientHeight;
-        var travelH = trackH - thumbH;
-
-        var trackY = 0;
-        var el = elTrack;
-        do {
-            trackY += el.offsetTop;
-        } while (el = el.offsetParent);
-
-        var offsetY = Math.max(0, Math.min(travelH, event.pageY - mouseCaptureOffset - trackY));
-        var value = 1.0 - offsetY / travelH;
-        elThumb.style.top = travelH * (1.0 - value) + 'px';
-    } else {
-        var thumbW = elThumb.clientWidth;
-        var trackW = elTrack.clientWidth;
-        var travelW = trackW - thumbW;
-
-        var trackX = 0;
-        var el = elTrack;
-        do {
-            trackX += el.offsetLeft;
-        } while (el = el.offsetParent);
-
-        var offsetX = Math.max(0, Math.min(travelW, event.pageX - mouseCaptureOffset - trackX));
-        var value = offsetX / travelW;
-        elThumb.style.left = travelW * value + 'px';
-    }
-
-    sliderSetValue(mouseCapture, value);
-}
-
-function handleMouseUp() {
-    mouseCapture = null;
-}
-
-function sliderSetValue(slider, value) {
-    var pitchRate = Math.pow(2.0, 2.0 * (value - 0.5));
-
-    switch(slider) {
-    case 'effect_thumb':
-        // Change the volume of the convolution effect.
-        beatMod.setBeatEffectMix(value);
-        setEffectLevel(beatMod.theBeat);
-        break;
-    case 'kick_thumb':
-        beatMod.setBeatKickPitchVal(value);
-        kickPitch = pitchRate;
-        break;
-    case 'snare_thumb':
-        beatMod.setBeatSnarePitchVal(value);
-        snarePitch = pitchRate;
-        break;
-    case 'hihat_thumb':
-        beatMod.setBeatHihatPitchVal(value);
-        hihatPitch = pitchRate;
-        break;
-    case 'tom1_thumb':
-        beatMod.setBeatTom1PitchVal(value);
-        tom1Pitch = pitchRate;
-        break;
-    case 'tom2_thumb':
-        beatMod.setBeatTom2PitchVal(value);
-        tom2Pitch = pitchRate;
-        break;
-    case 'tom3_thumb':
-        beatMod.setBeatTom3PitchVal(value);
-        tom3Pitch = pitchRate;
-        break;
-    case 'swing_thumb':
-        beatMod.setBeatSwingFactor(value);
-        break;
-    }
-}
-
-function sliderSetPosition(slider, value) {
-    var elThumb = document.getElementById(slider);
-    var elTrack = elThumb.parentNode;
-
-    if (slider == 'swing_thumb') {
-        var thumbW = elThumb.clientWidth;
-        var trackW = elTrack.clientWidth;
-        var travelW = trackW - thumbW;
-
-        elThumb.style.left = travelW * value + 'px';
-    } else {
-        var thumbH = elThumb.clientHeight;
-        var trackH = elTrack.clientHeight;
-        var travelH = trackH - thumbH;
-
-        elThumb.style.top = travelH * (1.0 - value) + 'px';
-    }
-}
-
-function handleButtonMouseDown(event) {
-    var notes = beatMod.theBeat.rhythm1;
-
-    var instrumentIndex;
-    var rhythmIndex;
-
-    var elId = event.target.id;
-    rhythmIndex = elId.substr(elId.indexOf('_') + 1, 2);
-    instrumentIndex = kitMod.instruments.indexOf(elId.substr(0, elId.indexOf('_')));
-
-    switch (instrumentIndex) {
-        case 0: notes = beatMod.theBeat.rhythm1; break;
-        case 1: notes = beatMod.theBeat.rhythm2; break;
-        case 2: notes = beatMod.theBeat.rhythm3; break;
-        case 3: notes = beatMod.theBeat.rhythm4; break;
-        case 4: notes = beatMod.theBeat.rhythm5; break;
-        case 5: notes = beatMod.theBeat.rhythm6; break;
-    }
-
-    var newNoteValue = (notes[rhythmIndex] + 1) % 3;
-
-    notes[rhythmIndex] = newNoteValue
-
-    if (instrumentIndex == currentlyActiveInstrument)
-        showCorrectNote( rhythmIndex, notes[rhythmIndex] );
-
-    drawMod.drawNote(notes[rhythmIndex], rhythmIndex, instrumentIndex);
-
-    if (newNoteValue) {
-        switch(instrumentIndex) {
-        case 0:  // Kick
-          playNote(kitMod.currentKit.kickBuffer, false, 0,0,-2, 0.5 * beatMod.theBeat.effectMix, kitMod.volumes[newNoteValue] * 1.0, kickPitch, 0);
-          break;
-
-        case 1:  // Snare
-          playNote(kitMod.currentKit.snareBuffer, false, 0,0,-2, beatMod.theBeat.effectMix, kitMod.volumes[newNoteValue] * 0.6, snarePitch, 0);
-          break;
-
-        case 2:  // Hihat
-          // Pan the hihat according to sequence position.
-          playNote(kitMod.currentKit.hihatBuffer, true, 0.5*rhythmIndex - 4, 0, -1.0, beatMod.theBeat.effectMix, kitMod.volumes[newNoteValue] * 0.7, hihatPitch, 0);
-          break;
-
-        case 3:  // Tom 1
-          playNote(kitMod.currentKit.tom1, false, 0,0,-2, beatMod.theBeat.effectMix, kitMod.volumes[newNoteValue] * 0.6, tom1Pitch, 0);
-          break;
-
-        case 4:  // Tom 2
-          playNote(kitMod.currentKit.tom2, false, 0,0,-2, beatMod.theBeat.effectMix, kitMod.volumes[newNoteValue] * 0.6, tom2Pitch, 0);
-          break;
-
-        case 5:  // Tom 3
-          playNote(kitMod.currentKit.tom3, false, 0,0,-2, beatMod.theBeat.effectMix, kitMod.volumes[newNoteValue] * 0.6, tom3Pitch, 0);
-          break;
-        }
-    }
-
-    synthMod.synthCode(newNoteValue, rhythmIndex, instrumentIndex)
-}
-
-function handleKitComboMouseDown(event) {
-    document.getElementById('kitcombo').classList.toggle('active');
-}
-
-function handleKitMouseDown(event) {
-    var index = kitMod.kitNamePretty.indexOf(event.target.innerHTML);
-    beatMod.setBeatKitIndex(index);
-    kitMod.setCurrentKit(kitMod.kits[index]);
-    document.getElementById('kitname').innerHTML = kitMod.kitNamePretty[index];
-}
-
-function handleBodyMouseDown(event) {
-    var elKitcombo = document.getElementById('kitcombo');
-    var elEffectcombo = document.getElementById('effectcombo');
-
-    if (elKitcombo.classList.contains('active') && !isDescendantOfId(event.target, 'kitcombo_container')) {
-        elKitcombo.classList.remove('active');
-        if (!isDescendantOfId(event.target, 'effectcombo_container')) {
-            event.stopPropagation();
-        }
-    }
-
-    if (elEffectcombo.classList.contains('active') && !isDescendantOfId(event.target, 'effectcombo')) {
-        elEffectcombo.classList.remove('active');
-        if (!isDescendantOfId(event.target, 'kitcombo_container')) {
-            event.stopPropagation();
-        }
-    }
-}
-
-function isDescendantOfId(el, id) {
-    if (el.parentElement) {
-        if (el.parentElement.id == id) {
-            return true;
-        } else {
-            return isDescendantOfId(el.parentElement, id);
-        }
-    } else {
-        return false;
-    }
-}
-
-function handleEffectComboMouseDown(event) {
-    if (event.target.id != 'effectlist') {
-        document.getElementById('effectcombo').classList.toggle('active');
-    }
-}
-
-function handleEffectMouseDown(event) {
-    for (var i = 0; i < impulseMod.impulseResponseInfoList.length; ++i) {
-        if (impulseMod.impulseResponseInfoList[i].name == event.target.innerHTML) {
-
-            // Hack - if effect is turned all the way down - turn up effect slider.
-            // ... since they just explicitly chose an effect from the list.
-            if (beatMod.theBeat.effectMix == 0)
-                beatMod.setBeatEffectMix(0.5);
-
-            setEffect(i);
-            break;
-        }
-    }
-}
 
 function setEffect(index) {
     if (index > 0 && !impulseMod.impulseResponseList[index].isLoaded()) {
@@ -693,183 +314,58 @@ function setEffect(index) {
     }
 
     beatMod.setBeatEffectIndex(index)
-    effectDryMix = impulseMod.impulseResponseInfoList[index].dryMix;
-    effectWetMix = impulseMod.impulseResponseInfoList[index].wetMix;
+    impulseMod.setEffectDryMix(impulseMod.impulseResponseInfoList[index].dryMix);
+    impulseMod.setEffectWetMix(impulseMod.impulseResponseInfoList[index].wetMix);
     convolver.buffer = impulseMod.impulseResponseList[index].buffer;
 
   // Hack - if the effect is meant to be entirely wet (not unprocessed signal)
   // then put the effect level all the way up.
-    if (effectDryMix == 0)
+    if (impulseMod.effectDryMix == 0)
         beatMod.setBeatEffectMix(1);
 
     setEffectLevel(beatMod.theBeat);
-    sliderSetValue('effect_thumb', beatMod.theBeat.effectMix);
+    handlersMod.sliderSetValue('effect_thumb', beatMod.theBeat.effectMix);
     updateControls();
 
     document.getElementById('effectname').innerHTML = impulseMod.impulseResponseInfoList[index].name;
 }
+exports.setEffect = setEffect;
 
 function setEffectLevel() {
     // Factor in both the preset's effect level and the blending level (effectWetMix) stored in the effect itself.
-    effectLevelNode.gain.value = beatMod.theBeat.effectMix * effectWetMix;
+    effectLevelNode.gain.value = beatMod.theBeat.effectMix * impulseMod.effectWetMix;
+    exports.effectLevelNode = effectLevelNode;
 }
+exports.setEffectLevel = setEffectLevel;
 
-
-function handleDemoMouseDown(event) {
-    var loaded = false;
-
-    switch(event.target.id) {
-        case 'demo1':
-            loaded = loadBeat(beatMod.beatDemo[0]);
-            break;
-        case 'demo2':
-            loaded = loadBeat(beatMod.beatDemo[1]);
-            break;
-        case 'demo3':
-            loaded = loadBeat(beatMod.beatDemo[2]);
-            break;
-        case 'demo4':
-            loaded = loadBeat(beatMod.beatDemo[3]);
-            break;
-        case 'demo5':
-            loaded = loadBeat(beatMod.beatDemo[4]);
-            break;
-    }
-
-    if (loaded)
-        handlePlay();
-}
-
-function handlePlay(event) {
-    noteTime = 0.0;
-    startTime = context.currentTime + 0.005;
-    schedule();
-    timerWorker.postMessage("start");
-
-    document.getElementById('play').classList.add('playing');
-    document.getElementById('stop').classList.add('playing');
-    if (midiOut) {
-        // turn off the play button
-        midiOut.send( [0x80, 3, 32] );
-        // light up the stop button
-        midiOut.send( [0x90, 7, 1] );
-    }
-}
-
-function handleStop(event) {
-    timerWorker.postMessage("stop");
-
-    var elOld = document.getElementById('LED_' + (beatMod.rhythmIndex + 14) % 16);
-    elOld.src = 'images/LED_off.png';
-
-    hideBeat( (beatMod.rhythmIndex + 14) % 16 );
-
-    beatMod.setRhythmIndex(0);
-
-    document.getElementById('play').classList.remove('playing');
-    document.getElementById('stop').classList.remove('playing');
-    if (midiOut) {
-        // light up the play button
-        midiOut.send( [0x90, 3, 32] );
-        // turn off the stop button
-        midiOut.send( [0x80, 7, 1] );
-    }
-}
-
-function handleSave(event) {
-    toggleSaveContainer();
-    var elTextarea = document.getElementById('save_textarea');
-    elTextarea.value = JSON.stringify(beatMod.theBeat);
-}
-
-function handleSaveOk(event) {
-    toggleSaveContainer();
-}
-
-function handleLoad(event) {
-    toggleLoadContainer();
-}
-
-function handleLoadOk(event) {
-    var elTextarea = document.getElementById('load_textarea');
-    beatMod.setBeat(JSON.parse(elTextarea.value));
-
-    // Set drumkit
-    kitMod.setCurrentKit(kitMod.kits[beatMod.theBeat.kitIndex]);
-    document.getElementById('kitname').innerHTML = kitMod.kitNamePretty[beatMod.theBeat.kitIndex];
-
-    // Set effect
-    setEffect(beatMod.theBeat.effectIndex);
-
-    // Change the volume of the convolution effect.
-    setEffectLevel(beatMod.theBeat);
-
-    // Apply values from sliders
-    sliderSetValue('effect_thumb', beatMod.theBeat.effectMix);
-    sliderSetValue('kick_thumb', beatMod.theBeat.kickPitchVal);
-    sliderSetValue('snare_thumb', beatMod.theBeat.snarePitchVal);
-    sliderSetValue('hihat_thumb', beatMod.theBeat.hihatPitchVal);
-    sliderSetValue('tom1_thumb', beatMod.theBeat.tom1PitchVal);
-    sliderSetValue('tom2_thumb', beatMod.theBeat.tom2PitchVal);
-    sliderSetValue('tom3_thumb', beatMod.theBeat.tom3PitchVal);
-    sliderSetValue('swing_thumb', beatMod.theBeat.swingFactor);
-
-    // Clear out the text area post-processing
-    elTextarea.value = '';
-
-    toggleLoadContainer();
-    updateControls();
-}
-
-function handleLoadCancel(event) {
-    toggleLoadContainer();
-}
-
-function toggleSaveContainer() {
-    document.getElementById('pad').classList.toggle('active');
-    document.getElementById('params').classList.toggle('active');
-    document.getElementById('tools').classList.toggle('active');
-    document.getElementById('save_container').classList.toggle('active');
-}
-
-function toggleLoadContainer() {
-    document.getElementById('pad').classList.toggle('active');
-    document.getElementById('params').classList.toggle('active');
-    document.getElementById('tools').classList.toggle('active');
-    document.getElementById('load_container').classList.toggle('active');
-}
-
-function handleReset(event) {
-    handleStop();
-    loadBeat(beatMod.beatReset);
-}
 
 function loadBeat(beat) {
     // Check that assets are loaded.
     if (beat != beatMod.beatReset && !beat.isLoaded())
         return false;
 
-    handleStop();
+    handlersMod.handleStop();
 
     beatMod.setBeat(beatMod.cloneBeat(beat));
     kitMod.setCurrentKit(kitMod.kits[beatMod.theBeat.kitIndex]);
     setEffect(beatMod.theBeat.effectIndex);
 
     // apply values from sliders
-    sliderSetValue('effect_thumb', beatMod.theBeat.effectMix);
-    sliderSetValue('kick_thumb', beatMod.theBeat.kickPitchVal);
-    sliderSetValue('snare_thumb', beatMod.theBeat.snarePitchVal);
-    sliderSetValue('hihat_thumb', beatMod.theBeat.hihatPitchVal);
-    sliderSetValue('tom1_thumb', beatMod.theBeat.tom1PitchVal);
-    sliderSetValue('tom2_thumb', beatMod.theBeat.tom2PitchVal);
-    sliderSetValue('tom3_thumb', beatMod.theBeat.tom3PitchVal);
-    sliderSetValue('swing_thumb', beatMod.theBeat.swingFactor);
+    handlersMod.sliderSetValue('effect_thumb', beatMod.theBeat.effectMix);
+    handlersMod.sliderSetValue('kick_thumb', beatMod.theBeat.kickPitchVal);
+    handlersMod.sliderSetValue('snare_thumb', beatMod.theBeat.snarePitchVal);
+    handlersMod.sliderSetValue('hihat_thumb', beatMod.theBeat.hihatPitchVal);
+    handlersMod.sliderSetValue('tom1_thumb', beatMod.theBeat.tom1PitchVal);
+    handlersMod.sliderSetValue('tom2_thumb', beatMod.theBeat.tom2PitchVal);
+    handlersMod.sliderSetValue('tom3_thumb', beatMod.theBeat.tom3PitchVal);
+    handlersMod.sliderSetValue('swing_thumb', beatMod.theBeat.swingFactor);
 
     updateControls();
     setActiveInstrument(0);
 
     return true;
 }
+exports.loadBeat = loadBeat;
 
 function updateControls() {
     for (i = 0; i < beatMod.loopLength; ++i) {
@@ -890,14 +386,14 @@ function updateControls() {
     document.getElementById('kitname').innerHTML = kitMod.kitNamePretty[beatMod.theBeat.kitIndex];
     document.getElementById('effectname').innerHTML = impulseMod.impulseResponseInfoList[beatMod.theBeat.effectIndex].name;
     document.getElementById('tempo').innerHTML = beatMod.theBeat.tempo;
-    sliderSetPosition('swing_thumb', beatMod.theBeat.swingFactor);
-    sliderSetPosition('effect_thumb', beatMod.theBeat.effectMix);
-    sliderSetPosition('kick_thumb', beatMod.theBeat.kickPitchVal);
-    sliderSetPosition('snare_thumb', beatMod.theBeat.snarePitchVal);
-    sliderSetPosition('hihat_thumb', beatMod.theBeat.hihatPitchVal);
-    sliderSetPosition('tom1_thumb', beatMod.theBeat.tom1PitchVal);
-    sliderSetPosition('tom2_thumb', beatMod.theBeat.tom2PitchVal);
-    sliderSetPosition('tom3_thumb', beatMod.theBeat.tom3PitchVal);
+    drawMod.sliderSetPosition('swing_thumb', beatMod.theBeat.swingFactor);
+    drawMod.sliderSetPosition('effect_thumb', beatMod.theBeat.effectMix);
+    drawMod.sliderSetPosition('kick_thumb', beatMod.theBeat.kickPitchVal);
+    drawMod.sliderSetPosition('snare_thumb', beatMod.theBeat.snarePitchVal);
+    drawMod.sliderSetPosition('hihat_thumb', beatMod.theBeat.hihatPitchVal);
+    drawMod.sliderSetPosition('tom1_thumb', beatMod.theBeat.tom1PitchVal);
+    drawMod.sliderSetPosition('tom2_thumb', beatMod.theBeat.tom2PitchVal);
+    drawMod.sliderSetPosition('tom3_thumb', beatMod.theBeat.tom3PitchVal);
 }
 
 
@@ -913,11 +409,41 @@ function filterFrequencyFromCutoff( cutoff ) {
 }
 
 function setFilterCutoff( cutoff ) {
-    if (filterNode)
+    if (filterNode){
         filterNode.frequency.value = filterFrequencyFromCutoff( cutoff );
+        exports.filterNode = filterNode;
+    }
 }
 
 function setFilterQ( Q ) {
-    if (filterNode)
+    if (filterNode){
         filterNode.Q.value = Q;
+        exports.filterNode = filterNode;
+    }
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+//
