+++ public/js/impulse.js
@@ -40,6 +40,21 @@ function setImpulseResponseList (lst) {
 // }
 
 
+// Each effect impulse response has a specific overall desired dry and wet volume.
+// For example in the telephone filter, it's necessary to make the dry volume 0 to correctly hear the effect.
+var effectDryMix = 1.0;
+var effectWetMix = 1.0;
+
+function setEffectDryMix(m) {
+  effectDryMix = m;
+  exports.effectDryMix = effectDryMix;
+}
+
+function setEffectWetMix(m) {
+  effectWetMix = m;
+  exports.effectWetMix = effectWetMix;
+}
+
 
 function ImpulseResponse(url, index) {
     this.url = url;
@@ -97,11 +112,14 @@ exports.ImpulseResponse = ImpulseResponse;
 
 // functions
 exports.setImpulseResponseList = setImpulseResponseList;
+exports.setEffectDryMix = setEffectDryMix;
+exports.setEffectWetMix = setEffectWetMix;
 
 // variables
 exports.impulseResponseList = impulseResponseList;
 exports.impulseResponseInfoList = impulseResponseInfoList;
-
+exports.effectDryMix = effectDryMix;
+exports.effectWetMix = effectWetMix;
